🛠️refactor: make util function
